feat(AddNewContact): Update user infos on contact addition
